fastapi-nacos/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.2
nacos-sdk-python>=0.1.1
python-dotenv>=1.0.0